Removed Pane folder, added Create and Form components
