--- app/javascript/src/components/Dashboard/Notes/Pane/CreateNote.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-import { Pane, Typography } from "neetoui/v2";
-
-import formInitialValues from "constants/formInitialValues";
-
-import Form from "./Form";
-
-export default function NewNotePane({ showPane, setShowPane }) {
-  const onClose = () => setShowPane(false);
-
-  return (
-    <Pane isOpen={showPane} onClose={onClose}>
-      <Pane.Header>
-        <Typography style="h2" weight="semibold">
-          Create a New Note
-        </Typography>
-      </Pane.Header>
-      <Form onClose={onClose} note={formInitialValues.notesForm} />
-    </Pane>
-  );
-}
