--- app/javascript/src/components/Dashboard/Notes/Pane/EditNote.jsx
@@ -1,25 +1,0 @@
-import React from "react";
-
-import { Pane, Typography } from "neetoui/v2";
-
-import Form from "./Form";
-
-export default function EditNotePane({
-  fetchNotes,
-  showPane,
-  setShowPane,
-  note,
-}) {
-  const onClose = () => setShowPane(false);
-
-  return (
-    <Pane isOpen={showPane} onClose={onClose}>
-      <Pane.Header>
-        <Typography style="h2" weight="semibold">
-          Edit Note
-        </Typography>
-      </Pane.Header>
-      <Form onClose={onClose} refetch={fetchNotes} note={note} isEdit={true} />
-    </Pane>
-  );
-}
