--- app/javascript/src/components/Dashboard/Notes/Pane/constants.js
@@ -1,8 +1,0 @@
-export const TAGS = [
-  "Getting Started",
-  "Onboarding",
-  "User Flow",
-  "UX",
-  "Bugs",
-  "V2",
-];
