+++ app/javascript/src/components/Dashboard/Notes/constants.js
@@ -32,3 +32,12 @@ export const NOTES = [
     assigned: "Oliver Smit",
   },
 ];
+
+export const TAGS = [
+  "Getting Started",
+  "Onboarding",
+  "User Flow",
+  "UX",
+  "Bugs",
+  "V2",
+];
